# Conflicts:
#	models/Asset.php

+++ models/Asset.php
@@ -54,6 +54,7 @@ use Symfony\Component\EventDispatcher\GenericEvent;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Lock\LockFactory;
 use Symfony\Component\Lock\LockInterface;
+use Throwable;
 
 /**
  * @method Dao getDao()
@@ -1657,8 +1658,12 @@ class Asset extends Element\AbstractElement
         try {
             $movedFiles = [];
             $children = $storage->listContents($oldPath, true);
-            foreach ($children as $child) {
-                if ($child['type'] === 'file') {
+            $totalChildren = iterator_count($children);
+
+            if ($totalChildren > 0) {
+                /** @var \League\Flysystem\StorageAttributes $child */
+                foreach ($children as $child) {
+                    if ($child instanceof \League\Flysystem\FileAttributes) {
                     $src  = $child['path'];
                     $dest = str_replace($oldPath, $newPath, '/' . $src);
 
@@ -1667,11 +1672,27 @@ class Asset extends Element\AbstractElement
                 }
             }
 
-            $storage->deleteDirectory($oldPath);
-        } catch (UnableToMoveFile $e) {
+                $movedCount = count($movedFiles);
+
+                if ($movedCount === $totalChildren) {
+                    $storage->deleteDirectory($oldPath);
+                } else {
+                    \Pimcore\Logger::info(
+                        sprintf(
+                            'Moved %d/%d files from %s to %s. No exception was thrown for %d files,
+                            so the source directory was not deleted.',
+                            $movedCount, $totalChildren, $oldPath, $newPath, $totalChildren - $movedCount
+                        )
+                    );
+                }
+            }
+        } catch (Throwable $e) {
+            Logger::error(sprintf('Asset Move to %s failed: %s', $newPath, $e->getMessage()));
+
             if ($skipError) {
                 return;
             }
+
             // rollback moved files
             foreach ($movedFiles as $src => $dest) {
                 $storage->move($src, $dest);
